requirements.txt
streamlit
requests
google-generativeai
openai>=1.37.0
